# api-docs/openapi-v3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "办公室绿萝 Pro API",
    "description": "办公室绿萝生长监测系统 v3.0 接口文档",
    "version": "3.0.0"
  },
  "servers": [
    {"url": "http://localhost:5501/api", "description": "本地开发环境"}
  ],
  "paths": {
...
    "/detect": {
      "post": {
        "summary": "执行目标检测",
        "description": "对指定图片进行YOLO目标检测",
        "tags": ["检测识别"],
        "parameters": [
          {"name": "imagePath", "in": "query", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "检测成功"},
          "400": {"description": "检测失败"}
        }
      }
    },
    "/detect/batch": {
      "post": {
        "summary": "批量目标检测",
        "description": "对多张图片进行批量YOLO目标检测",
        "tags": ["检测识别"],
        "parameters": [
          {"name": "imagePaths", "in": "query", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "批量检测完成"}
        }
      }
    },
...
  },
...
}